+++ src/main/java/frc/robot/AutoConstants.java
@@ -1,4 +1,4 @@
-package frc.robot.constants;
+package frc.robot;
 
 
 import com.pathplanner.lib.path.PathConstraints;
