+++ src/main/java/frc/robot/RobotCommands.java
@@ -4,9 +4,6 @@ import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.StartEndCommand;
-import frc.robot.constants.ArmConstants;
-import frc.robot.constants.PivotConstants;
-import frc.robot.constants.ShooterConstants;
 import frc.robot.subsystems.*;
 import frc.robot.subsystems.drivetrain.CommandSwerveDrivetrain;
 import frc.robot.util.RobotDataPublisher;
@@ -44,7 +41,7 @@ public class RobotCommands {
             final Pose2d speakerPose = AllianceFlipUtil.apply(new Pose2d(0.0, 5.5, Rotation2d.fromDegrees(0)));
             Pose2d relativeSpeaker = robotPose.relativeTo(speakerPose);
             double distance = relativeSpeaker.getTranslation().getNorm();
-            return ShooterConstants.SHOOTER_MAP().get(distance);
+            return Constants.ShooterConstants.SHOOTER_MAP().get(distance);
         });
 
         return pivot.updatePivotPositionWith(shooterStatePublisher)
@@ -52,9 +49,9 @@ public class RobotCommands {
     }
 
     public Command handoff() {
-        return pivot.movePivotPositionTo(PivotConstants.PivotPosition.HANDOFF).andThen(
+        return pivot.movePivotPositionTo(Constants.PivotConstants.PivotPosition.HANDOFF).andThen(
                 new StartEndCommand(() -> arm.moveUp(.5), arm::stop).until(() ->
-                        arm.getEnc() <= ArmConstants.ARM_HANDOFF_POSITION).andThen(
+                        arm.getEnc() <= Constants.ArmConstants.ARM_HANDOFF_POSITION).andThen(
                         shooter.spinFeederAndStop(-0.3).alongWith(intake.rollOut(-0.35))
                 ).until(shooter::getBeamBreak).andThen(intake.rollOut(1).withTimeout(0.2))
         );
