--- src/main/java/frc/robot/constants/TalonFXConstants.java
@@ -1,8 +1,0 @@
-package frc.robot.constants;
-
-public final class TalonFXConstants {
-    public static final int COUNTS_PER_REV = 2048;
-    public static final double COUNTS_PER_DEG = COUNTS_PER_REV / 360.0;
-    public static final boolean TALON_FUTURE_PROOF = true;
-    public static final String CANIVORE_NAME = "canivoreBus";
-}
