--- src/main/java/frc/robot/constants/VisionConstants.java
@@ -1,7 +1,0 @@
-package frc.robot.constants;
-
-public final class VisionConstants {
-    public static final String LIMELIGHT_NAME = "limelight";
-    public static final double BLINK_DELAY_SECONDS = 0.125;
-    public static final int NUM_BLINKS = 4;
-}
