+++ src/main/java/frc/robot/subsystems/ArmSubsystem.java
@@ -3,30 +3,23 @@ package frc.robot.subsystems;
 
 import com.ctre.phoenix6.configs.CANcoderConfiguration;
 import com.ctre.phoenix6.configs.TalonFXConfiguration;
-import com.ctre.phoenix6.controls.MotionMagicTorqueCurrentFOC;
-import com.ctre.phoenix6.controls.MotionMagicVelocityDutyCycle;
-import com.ctre.phoenix6.controls.MotionMagicVelocityVoltage;
 import com.ctre.phoenix6.controls.MotionMagicVoltage;
 import com.ctre.phoenix6.hardware.CANcoder;
 import com.ctre.phoenix6.hardware.TalonFX;
 import com.ctre.phoenix6.signals.*;
-import edu.wpi.first.wpilibj.DigitalInput;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.constants.ArmConstants;
-import frc.robot.constants.CANCoderConstants;
-import frc.robot.constants.TalonFXConstants;
-import org.opencv.core.Mat;
+import frc.robot.Constants;
 
 public class ArmSubsystem extends SubsystemBase {
 
     private final TalonFX armKraken;
     private final CANcoder armEncoder;
-    private ArmConstants.ArmPositions armPositions = ArmConstants.ArmPositions.STOWED;
+    private Constants.ArmConstants.ArmPositions armPositions = Constants.ArmConstants.ArmPositions.STOWED;
 
     public ArmSubsystem() {
-        armKraken = new TalonFX(ArmConstants.ARM_KRAKEN_ID, TalonFXConstants.CANIVORE_NAME);
-        armEncoder = new CANcoder(ArmConstants.ARM_CANCODER_ID, TalonFXConstants.CANIVORE_NAME);
+        armKraken = new TalonFX(Constants.ArmConstants.ARM_KRAKEN_ID, Constants.TalonFXConstants.CANIVORE_NAME);
+        armEncoder = new CANcoder(Constants.ArmConstants.ARM_CANCODER_ID, Constants.TalonFXConstants.CANIVORE_NAME);
 
         var armConfigurator = armKraken.getConfigurator();
         var talonFXConfiguration = new TalonFXConfiguration();
@@ -34,17 +27,17 @@ public class ArmSubsystem extends SubsystemBase {
 
         talonFXConfiguration.Feedback.FeedbackRemoteSensorID = armEncoder.getDeviceID();
         talonFXConfiguration.Feedback.FeedbackSensorSource = FeedbackSensorSourceValue.RemoteCANcoder;
-        talonFXConfiguration.MotorOutput.Inverted = ArmConstants.ARM_INVERSION;
-        talonFXConfiguration.MotorOutput.NeutralMode = ArmConstants.ARM_NEUTRAL_MODE;
-        talonFXConfiguration.FutureProofConfigs = TalonFXConstants.TALON_FUTURE_PROOF;
+        talonFXConfiguration.MotorOutput.Inverted = Constants.ArmConstants.ARM_INVERSION;
+        talonFXConfiguration.MotorOutput.NeutralMode = Constants.ArmConstants.ARM_NEUTRAL_MODE;
+        talonFXConfiguration.FutureProofConfigs = Constants.TalonFXConstants.TALON_FUTURE_PROOF;
         talonFXConfiguration.Feedback.SensorToMechanismRatio = 50; //placeholder
         talonFXConfiguration.Feedback.RotorToSensorRatio = 12.8;
-        talonFXConfiguration.CurrentLimits = ArmConstants.ARM_CURRENT_LIMIT;
-        talonFXConfiguration.SoftwareLimitSwitch = ArmConstants.ARM_SOFT_LIMIT;
-        talonFXConfiguration.Slot0 = ArmConstants.SLOT_0_CONFIGS;
+        talonFXConfiguration.CurrentLimits = Constants.ArmConstants.ARM_CURRENT_LIMIT;
+        talonFXConfiguration.SoftwareLimitSwitch = Constants.ArmConstants.ARM_SOFT_LIMIT;
+        talonFXConfiguration.Slot0 = Constants.ArmConstants.SLOT_0_CONFIGS;
 
-        armKraken.getRotorVelocity().waitForUpdate(ArmConstants.ARM_VELOCITY_STATUS_FRAME);
-        armKraken.getRotorPosition().waitForUpdate(ArmConstants.ARM_POSITION_STATUS_FRAME);
+        armKraken.getRotorVelocity().waitForUpdate(Constants.ArmConstants.ARM_VELOCITY_STATUS_FRAME);
+        armKraken.getRotorPosition().waitForUpdate(Constants.ArmConstants.ARM_POSITION_STATUS_FRAME);
 
         armConfigurator.apply(talonFXConfiguration);
 
@@ -52,7 +45,7 @@ public class ArmSubsystem extends SubsystemBase {
         var cancoderConfiguration = new CANcoderConfiguration();
 
         cancoderConfiguration.MagnetSensor.AbsoluteSensorRange = AbsoluteSensorRangeValue.Unsigned_0To1;
-        cancoderConfiguration.MagnetSensor.MagnetOffset = ArmConstants.MAGNET_OFFSET;
+        cancoderConfiguration.MagnetSensor.MagnetOffset = Constants.ArmConstants.MAGNET_OFFSET;
         cancoderConfiguration.MagnetSensor.SensorDirection = SensorDirectionValue.Clockwise_Positive;
         armEncoderConfigurator.apply(cancoderConfiguration);
 
@@ -64,7 +57,7 @@ public class ArmSubsystem extends SubsystemBase {
     }
 
 
-    public void setPosition(ArmConstants.ArmPositions position) {
+    public void setPosition(Constants.ArmConstants.ArmPositions position) {
         armPositions = position;
         setMotorsBrake();
 
@@ -72,7 +65,7 @@ public class ArmSubsystem extends SubsystemBase {
         double cosineScalar = Math.cos(radians);
 
         MotionMagicVoltage motionMagicVoltage = new MotionMagicVoltage(
-                position.sensorUnits, true, ArmConstants.GRAVITY_FEEDFORWARD * cosineScalar, 0,
+                position.sensorUnits, true, Constants.ArmConstants.GRAVITY_FEEDFORWARD * cosineScalar, 0,
                 true, false, false);
 
         armKraken.setControl(motionMagicVoltage);
@@ -82,10 +75,10 @@ public class ArmSubsystem extends SubsystemBase {
     }
 
     public double getAngle() {
-        return armKraken.getRotorPosition().getValue() / CANCoderConstants.COUNTS_PER_DEG * ArmConstants.GEAR_RATIO;
+        return armKraken.getRotorPosition().getValue() / Constants.CANCoderConstants.COUNTS_PER_DEG * Constants.ArmConstants.GEAR_RATIO;
     }
 
-    public ArmConstants.ArmPositions getArmPositions() {
+    public Constants.ArmConstants.ArmPositions getArmPositions() {
         return armPositions;
     }
 
@@ -94,7 +87,7 @@ public class ArmSubsystem extends SubsystemBase {
     }
 
     public boolean isMotionFinished() {
-        return Math.abs(getAngle() - armPositions.angle) <= ArmConstants.ANGLE_TOLERANCE;
+        return Math.abs(getAngle() - armPositions.angle) <= Constants.ArmConstants.ANGLE_TOLERANCE;
     }
 
     public void moveUp(double speed) {
