+++ src/main/java/frc/robot/subsystems/IntakeSubsystem.java
@@ -5,27 +5,24 @@ import com.ctre.phoenix6.hardware.TalonFX;
 import edu.wpi.first.wpilibj.DigitalInput;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
-import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import edu.wpi.first.wpilibj2.command.button.Trigger;
-import frc.robot.constants.IntakeConstants;
-import frc.robot.constants.TalonFXConstants;
+import frc.robot.Constants;
 
 public class IntakeSubsystem extends SubsystemBase {
     private final TalonFX intakeKraken;
     private final DigitalInput beamBreak;
 
     public IntakeSubsystem() {
-        intakeKraken = new TalonFX(IntakeConstants.INTAKE_KRAKEN_ID, "canivoreBus");
+        intakeKraken = new TalonFX(Constants.IntakeConstants.INTAKE_KRAKEN_ID, "canivoreBus");
         var intakeConfigurator = intakeKraken.getConfigurator();
         var configs = new TalonFXConfiguration();
 
         beamBreak = new DigitalInput(2);
-        configs.MotorOutput.Inverted = IntakeConstants.INTAKE_INVERSION;
-        configs.MotorOutput.NeutralMode = IntakeConstants.INTAKE_NEUTRAL_MODE;
-        configs.FutureProofConfigs = TalonFXConstants.TALON_FUTURE_PROOF;
-        intakeKraken.getRotorVelocity().waitForUpdate(IntakeConstants.INTAKE_VELOCITY_STATUS_FRAME);
-        intakeKraken.getRotorPosition().waitForUpdate(IntakeConstants.INTAKE_POSITION_STATUS_FRAME);
+        configs.MotorOutput.Inverted = Constants.IntakeConstants.INTAKE_INVERSION;
+        configs.MotorOutput.NeutralMode = Constants.IntakeConstants.INTAKE_NEUTRAL_MODE;
+        configs.FutureProofConfigs = Constants.TalonFXConstants.TALON_FUTURE_PROOF;
+        intakeKraken.getRotorVelocity().waitForUpdate(Constants.IntakeConstants.INTAKE_VELOCITY_STATUS_FRAME);
+        intakeKraken.getRotorPosition().waitForUpdate(Constants.IntakeConstants.INTAKE_POSITION_STATUS_FRAME);
         intakeConfigurator.apply(configs);
     }
